Combining uart serial comms into mav library
