+++ SquirrelDefender/CMakeLists.txt
@@ -161,13 +161,13 @@ endfunction()
 # Optionally include code that is specific to different platforms
 if (BLD_JETSON_B01)
     add_definitions(-DBLD_JETSON_B01)
-    add_definitions(-DBLD_JETSON_SERIAL)
+    add_definitions(-DBLD_LINUX_SERIAL)
     message("Building for BLD_JETSON_B01 with serial comms.")
     add_definitions(-DENABLE_CV)
     message("Enabled video and computer vision functionality.")
 elseif (BLD_JETSON_ORIN_NANO)
     add_definitions(-DBLD_JETSON_ORIN_NANO)
-    add_definitions(-DBLD_JETSON_SERIAL)
+    add_definitions(-DBLD_LINUX_SERIAL)
     message("Building for BLD_JETSON_ORIN_NANO with serial comms.")
     add_definitions(-DENABLE_CV)
     message("Enabled video and computer vision functionality.")
