+++ SquirrelDefender/appsrc/mav_data_hub.cpp
@@ -148,9 +148,9 @@ const int32_t MESSAGE_RATE_1Hz = 1000000;
 /********************************************************************************
  * Function definitions
  ********************************************************************************/
-bool start_mav_comm(void) { return Serial::start_uart_comm(); }; // Open up uart port for mavlink messages
-void stop_mav_comm(void) { Serial::stop_uart_comm(); };          // Stop mavlink comms on uart port
-uint8_t read_mav_msg(void) { return Serial::read_uart(); };      // Read a byte
+bool start_mav_comm(void) { return MavSerial::start_mav_comm(); }; // Open up uart port for mavlink messages
+void stop_mav_comm(void) { MavSerial::stop_mav_comm(); };          // Stop mavlink comms on uart port
+uint8_t read_mav_msg(void) { return MavSerial::read_mav_msg(); };      // Read a byte
 void subscribe(uint16_t msg_id, float msg_interval);             // Subscribe to a mavlink message at desired rate
 void set_msg_rate(uint16_t msg_id, float msg_interval) { MavCmd::set_mav_msg_rate(SENDER_SYS_ID, SENDER_COMP_ID, TARGET_SYS_ID, TARGET_COMP_ID, msg_id, msg_interval); };
 void req_msg(uint16_t msg_id) { MavCmd::req_mav_msg(SENDER_SYS_ID, SENDER_COMP_ID, TARGET_SYS_ID, TARGET_COMP_ID, msg_id); };
@@ -657,7 +657,7 @@ void parse_mav_msgs(void)
     mavlink_status_t status = {}; // Initialize the Mavlink status
     uint8_t byte;
 
-    int n = Serial::bytes_available();
+    int n = MavSerial::bytes_available();
 
     for (int i = n; i > 0; i--)
     {
