+++ SquirrelDefender/libraries/mav/mav_cmd.h
@@ -11,7 +11,7 @@
 /********************************************************************************
  * Includes
  ********************************************************************************/
-#include "serial.h"
+#include "mav_serial.h"
 #include <mavlink.h>
 #include <common.h>
 
@@ -52,10 +52,10 @@ public:
     static void send_cmd_int(uint8_t sender_sys_id, uint8_t sender_comp_id, const mavlink_command_int_t& command_int);
     
 private:
-    static void send_mav_cmd(mavlink_message_t &msg) { Serial::write_uart(msg); };
-    static bool start_mav_comm(void) { return Serial::start_uart_comm(); }; // Open up uart port for mavlink messages
-    static void stop_mav_comm(void) { Serial::stop_uart_comm(); };          // Stop mavlink comms on uart port
-    static uint8_t read_mav_msg(void) { return Serial::read_uart(); };      // Read a byte
+    static void send_mav_cmd(mavlink_message_t &msg) { MavSerial::write_mav_msg(msg); };
+    static bool start_mav_comm(void) { return MavSerial::start_mav_comm(); }; // Open up uart port for mavlink messages
+    static void stop_mav_comm(void) { MavSerial::stop_mav_comm(); };          // Stop mavlink comms on uart port
+    static uint8_t read_mav_msg(void) { return MavSerial::read_mav_msg(); };      // Read a byte
     static void subscribe(uint16_t msg_id, float msg_interval);                 // Subscribe to a mavlink message at desired rate
 };
