+++ SquirrelDefender/libraries/mav/mav_motion.cpp
