+++ SquirrelDefender/libraries/mav/mav_msg_print.cpp
